Query remembering
When clicking on save/follow it used to be that the page number and search was forgotten. This commit fixes that

Also moved shared methods to a super class called CheepPageModel, such that there isn't as much code duplication
