+++ src/Chirp.Web/Pages/Public.cshtml
@@ -4,7 +4,7 @@
     ViewData["Title"] = "Chirp!";
     Layout = "Shared/_Layout";
     int currentPage = 1;
-    var pageQuery = Model.HttpContext.Request.Query["page"];
+    var pageQuery = Model.HttpContext.Request.Query["pageIndex"];
     if (!string.IsNullOrEmpty(pageQuery))
     {
         int.TryParse(pageQuery, out currentPage);
@@ -14,7 +14,7 @@
 
 <div>
     <form method="post" asp-page-handler="Search">
-        <input style="float: left" type="text" asp-for="SearchText">
+        <input style="float: left" type="text" asp-for="Search">
         <input type="submit" value="Search">
     </form>
     <br/>
@@ -26,7 +26,7 @@
         <h2> Public Timeline </h2>
         <div class="cheepbox">
             <h3>What's on your mind @(User.Identity.Name)?</h3>
-            <form method="post" asp-page-handler="Cheep">
+            <form method="post" asp-page-handler="Cheep" asp-route-search="@Request.Query["search"]" asp-route-pageIndex="@Model.PageIndex">
                 <input style="float: left" type="text" asp-for="Text">
                 <input type="submit" value="Share">
             </form>
@@ -47,14 +47,14 @@
                                 var isFollowing = await Model.IsFollowingAsync(User.Identity.Name!, cheep.Author.Name);
                                 if (isFollowing && User.Identity.Name != cheep.Author.Name)
                                 {
-                                    <form method="post" asp-page-handler="Unfollow" style="display:inline;">
+                                    <form method="post" asp-page-handler="Unfollow" asp-route-search="@Request.Query["search"]" asp-route-pageIndex="@Model.PageIndex" style="display:inline;">
                                         <button type="submit" class="btn btn-link p-0" name="unfollow"
                                             value="@cheep.Author.Name">unfollow</button>
                                     </form>
                                 }
                                 else if (User.Identity.Name != cheep.Author.Name)
                                 {
-                                    <form method="post" asp-page-handler="Follow" style="display:inline;">
+                                    <form method="post" asp-page-handler="Follow" asp-route-search="@Request.Query["search"]" asp-route-pageIndex="@Model.PageIndex"  style="display:inline;">
                                         <button type="submit" class="btn btn-link p-0" name="follow"
                                             value="@cheep.Author.Name">follow</button>
                                     </form>
@@ -62,12 +62,12 @@
 
                                 var isSaved = await Model.IsSavedAsync(cheep);
                                 if (isSaved) {
-                                    <form method="post" asp-page-handler="RemoveSave" style="display:inline;">
+                                    <form method="post" asp-page-handler="RemoveSave" asp-route-search="@Request.Query["search"]" asp-route-pageIndex="@Model.PageIndex"  style="display:inline;">
                                         <button type="submit" class="btn btn-link p-0" name="unsave"
                                             value="@cheep.CheepId">unsave</button>
                                     </form>
                                 } else {
-                                    <form method="post" asp-page-handler="Save" style="display:inline;">
+                                    <form method="post" asp-page-handler="Save" asp-route-search="@Request.Query["search"]" asp-route-pageIndex="@Model.PageIndex"  style="display:inline;">
                                         <button type="submit" class="btn btn-link p-0" name="save"
                                             value="@cheep.CheepId">save</button>
                                     </form>
@@ -84,11 +84,11 @@
         <div class="pagination">
             @if (currentPage > 1)
             {
-                <a href="/?page=@(currentPage - 1)&search=@Request.Query["search"]">Prev</a>
+                <a href="/?pageIndex=@(currentPage - 1)&search=@Request.Query["search"]">Prev</a>
             }
             <span>Page @currentPage</span>
 
-            <a id="next" href="/?page=@(currentPage + 1)&search=@Request.Query["search"]">Next</a>
+            <a id="next" href="/?pageIndex=@(currentPage + 1)&search=@Request.Query["search"]">Next</a>
         </div>
     }
     else
@@ -97,7 +97,7 @@
         <br/>
         @if (currentPage > 1)
         {
-            <a href="/?page=@(currentPage - 1)&search=@Request.Query["search"]">Prev</a>
+            <a href="/?pageIndex=@(currentPage - 1)&search=@Request.Query["search"]">Prev</a>
         }
         <span>Page @currentPage</span>
     }
