+++ src/Chirp.Web/Pages/Saved.cshtml
@@ -4,7 +4,7 @@
     ViewData["Title"] = "Chirp!";
     Layout = "Shared/_Layout";
     int currentPage = 1;
-    var pageQuery = Model.HttpContext.Request.Query["page"];
+    var pageQuery = Model.HttpContext.Request.Query["pageIndex"];
     if (!string.IsNullOrEmpty(pageQuery))
     {
         int.TryParse(pageQuery, out currentPage);
@@ -32,20 +32,20 @@
                                 var isFollowing = await Model.IsFollowingAsync(User.Identity.Name!, cheep.Author.Name);
                                 if (isFollowing && User.Identity.Name != cheep.Author.Name)
                                 {
-                                    <form method="post" asp-page-handler="Unfollow" style="display:inline;">
+                                    <form method="post" asp-page-handler="Unfollow" asp-route-search="@Request.Query["search"]" asp-route-pageIndex="@Model.PageIndex" style="display:inline;">
                                         <button type="submit" class="btn btn-link p-0" name="unfollow"
                                             value="@cheep.Author.Name">unfollow</button>
                                     </form>
                                 }
                                 else if (User.Identity.Name != cheep.Author.Name)
                                 {
-                                    <form method="post" asp-page-handler="Follow" style="display:inline;">
+                                    <form method="post" asp-page-handler="Follow" asp-route-search="@Request.Query["search"]" asp-route-pageIndex="@Model.PageIndex" style="display:inline;">
                                         <button type="submit" class="btn btn-link p-0" name="follow"
                                             value="@cheep.Author.Name">follow</button>
                                     </form>
                                 }
 
-                                <form method="post" asp-page-handler="RemoveSave" style="display:inline;">
+                                <form method="post" asp-page-handler="RemoveSave" asp-route-search="@Request.Query["search"]" asp-route-pageIndex="@Model.PageIndex" style="display:inline;">
                                     <button type="submit" class="btn btn-link p-0" name="unsave"
                                         value="@cheep.CheepId">unsave</button>
                                 </form>
@@ -60,11 +60,11 @@
         <div class="pagination">
             @if (currentPage > 1)
             {
-                <a href="/Saved/?page=@(currentPage - 1)&search=@Request.Query["search"]">Prev</a>
+                <a href="/Saved/?pageIndex=@(currentPage - 1)&search=@Request.Query["search"]">Prev</a>
             }
             <span>Page @currentPage</span>
 
-            <a href="/Saved/?page=@(currentPage + 1)&search=@Request.Query["search"]">Next</a>
+            <a href="/Saved/?pageIndex=@(currentPage + 1)&search=@Request.Query["search"]">Next</a>
         </div>
     }
     else
@@ -73,7 +73,7 @@
         <br/>
         @if (currentPage > 1)
         {
-            <a href="/Saved/?page=@(currentPage - 1)&search=@Request.Query["search"]">Prev</a>
+            <a href="/Saved/?pageIndex=@(currentPage - 1)&search=@Request.Query["search"]">Prev</a>
         }
         <span>Page @currentPage</span>
     }
